properly handle shop links in both OfferCard and AdvertisementCard
